refactor: Interceptor 교체
