--- src/main/java/nextstep/auth/authentication/SessionAuthenticationInterceptor2.java
@@ -1,27 +1,0 @@
-package nextstep.auth.authentication;
-
-import nextstep.auth.context.Authentication;
-import nextstep.auth.context.SecurityContext;
-import nextstep.member.application.CustomUserDetailsService;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
-import static nextstep.auth.context.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
-
-public class SessionAuthenticationInterceptor2 extends AuthenticationInterceptor {
-
-    public SessionAuthenticationInterceptor2(CustomUserDetailsService userDetailsService,
-                                             AuthenticationConverter converter) {
-        super(userDetailsService, converter);
-    }
-
-    @Override
-    public void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) {
-        HttpSession httpSession = request.getSession();
-        httpSession.setAttribute(SPRING_SECURITY_CONTEXT_KEY, new SecurityContext(authentication));
-        response.setStatus(HttpServletResponse.SC_OK);
-    }
-
-}
