--- src/main/java/nextstep/auth/authentication/TokenAuthenticationInterceptor2.java
@@ -1,40 +1,0 @@
-package nextstep.auth.authentication;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import nextstep.auth.context.Authentication;
-import nextstep.auth.token.JwtTokenProvider;
-import nextstep.auth.token.TokenResponse;
-import nextstep.member.application.CustomUserDetailsService;
-import org.springframework.http.MediaType;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-public class TokenAuthenticationInterceptor2 extends AuthenticationInterceptor {
-
-    private JwtTokenProvider jwtTokenProvider;
-    private ObjectMapper objectMapper;
-
-    public TokenAuthenticationInterceptor2(CustomUserDetailsService userDetailsService,
-                                           AuthenticationConverter converter,
-                                           JwtTokenProvider jwtTokenProvider,
-                                           ObjectMapper objectMapper) {
-        super(userDetailsService, converter);
-        this.jwtTokenProvider = jwtTokenProvider;
-        this.objectMapper = objectMapper;
-    }
-
-    @Override
-    public void afterAuthentication(HttpServletRequest request, HttpServletResponse response, Authentication authentication) throws IOException {
-        String payload = objectMapper.writeValueAsString(authentication.getPrincipal());
-        String token = jwtTokenProvider.createToken(payload);
-        TokenResponse tokenResponse = new TokenResponse(token);
-
-        String responseToClient = objectMapper.writeValueAsString(tokenResponse);
-        response.setStatus(HttpServletResponse.SC_OK);
-        response.setContentType(MediaType.APPLICATION_JSON_VALUE);
-        response.getOutputStream().print(responseToClient);
-    }
-
-}
