--- src/main/java/nextstep/auth/authorization/SessionSecurityContextPersistenceInterceptor2.java
@@ -1,16 +1,0 @@
-package nextstep.auth.authorization;
-
-import nextstep.auth.context.SecurityContext;
-
-import javax.servlet.http.HttpServletRequest;
-
-import static nextstep.auth.context.SecurityContextHolder.SPRING_SECURITY_CONTEXT_KEY;
-
-public class SessionSecurityContextPersistenceInterceptor2 extends SecurityContextInterceptor {
-
-    @Override
-    protected SecurityContext getSecurityContext(HttpServletRequest request) {
-        return (SecurityContext) request.getSession().getAttribute(SPRING_SECURITY_CONTEXT_KEY);
-    }
-
-}
