+++ src/test/java/nextstep/subway/unit/TokenAuthenticationInterceptorTest.java
@@ -1,9 +1,7 @@
 package nextstep.subway.unit;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import nextstep.auth.authentication.AuthenticationToken;
-import nextstep.auth.authentication.TokenAuthenticationInterceptor;
-import nextstep.auth.context.Authentication;
+import nextstep.auth.authentication.*;
 import nextstep.auth.token.JwtTokenProvider;
 import nextstep.auth.token.TokenRequest;
 import nextstep.auth.token.TokenResponse;
@@ -11,6 +9,7 @@ import nextstep.member.application.CustomUserDetailsService;
 import nextstep.member.domain.LoginMember;
 import org.apache.http.HttpStatus;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockHttpServletRequest;
@@ -24,6 +23,7 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+@DisplayName("토큰 기반 인증")
 class TokenAuthenticationInterceptorTest {
     private static final String EMAIL = "email@email.com";
     private static final String PASSWORD = "password";
@@ -31,48 +31,25 @@ class TokenAuthenticationInterceptorTest {
 
     private CustomUserDetailsService userDetailsService;
     private JwtTokenProvider jwtTokenProvider;
-    private TokenAuthenticationInterceptor interceptor;
     private ObjectMapper objectMapper;
+    private AuthenticationConverter converter;
+    private AuthenticationInterceptor interceptor;
     @BeforeEach
     void setUp() {
         userDetailsService = mock(CustomUserDetailsService.class);
         jwtTokenProvider = mock(JwtTokenProvider.class);
         objectMapper = new ObjectMapper();
-        interceptor = new TokenAuthenticationInterceptor(userDetailsService, jwtTokenProvider, objectMapper);
+        converter = mock(TokenAuthenticationConverter.class);
+        interceptor = new TokenAuthenticationInterceptor(userDetailsService, converter, jwtTokenProvider, objectMapper);
     }
 
     @Test
-    void convert() throws IOException {
-        //when
-        AuthenticationToken authenticationToken = interceptor.convert(createMockRequest());
-
-        //then
-        assertAll(
-                () -> assertThat(authenticationToken.getPrincipal()).isEqualTo(EMAIL),
-                () -> assertThat(authenticationToken.getCredentials()).isEqualTo(PASSWORD)
-        );
-    }
-
-    @Test
-    void authenticate() {
-        //given
-        LoginMember loginMember = new LoginMember(1L, EMAIL, PASSWORD, 20);
-        when(userDetailsService.loadUserByUsername(EMAIL)).thenReturn(loginMember);
-        AuthenticationToken authenticationToken = new AuthenticationToken(EMAIL, PASSWORD);
-
-        //when
-        Authentication authenticate = interceptor.authenticate(authenticationToken);
-
-        //then
-        assertThat(authenticate.getPrincipal()).isEqualTo(loginMember);
-    }
-
-    @Test
-    void preHandle() throws IOException {
+    void preHandle() throws Exception {
         //given
         LoginMember loginMember = new LoginMember(1L, EMAIL, PASSWORD, 20);
         when(userDetailsService.loadUserByUsername(EMAIL)).thenReturn(loginMember);
         when(jwtTokenProvider.createToken(any())).thenReturn(JWT_TOKEN);
+        when(converter.convert(any())).thenReturn(new AuthenticationToken(EMAIL, PASSWORD));
 
         MockHttpServletResponse response = new MockHttpServletResponse();
